:bug: fix prev/next cluttering page title
Fixed an issue with clicking prev/next buttons while a song is playing
which caused the play emoji added to the page title to be added again.

Fix is to call the playBack() function to pause the track when switching
which will also remove the play emoji from the page title appropriately.